+++ crates/agent/Cargo.toml
@@ -11,6 +11,7 @@ rust.unused_crate_dependencies = "deny"
 rust.unused_must_use = "deny"
 
 [dependencies]
+strata-bridge-connectors.workspace = true
 strata-bridge-db.workspace = true
 strata-bridge-primitives.workspace = true
 strata-bridge-proof-primitives.workspace = true
@@ -38,7 +39,7 @@ rkyv.workspace = true
 secp256k1 = { workspace = true, features = ["global-context", "rand-std"] }
 serde.workspace = true
 serde_json.workspace = true
-sp1-verifier = "4.0.0"
+sp1-verifier.workspace = true
 tokio.workspace = true
 tracing.workspace = true
