+++ crates/connectors/src/connector_a30.rs
@@ -1,3 +1,5 @@
+//! This module contains connector for the first output of the PostAssert transaction.
+// FIXME: remove this connector once the stake chain is integrated.
 use bitcoin::{
     psbt::Input,
     taproot::{self, ControlBlock, LeafVersion, TaprootSpendInfo},
@@ -24,7 +26,9 @@ pub struct ConnectorA30 {
 /// leaf is spent.
 #[derive(Debug, Clone, Copy)]
 pub enum ConnectorA30Leaf<Witness = ()> {
+    /// The leaf used in the Payout transaction.
     Payout(Witness),
+    /// The leaf used in the Disprove transaction.
     Disprove(Witness),
 }
