+++ crates/connectors/src/connector_p.rs
@@ -1,71 +1,22 @@
+//! This module contains the connector for the second output of the Stake transaction.
+//!
+//! This connector is used to either settle the payout transactions or to burn them.
 use bitcoin::{
     hashes::{sha256, Hash},
-    opcodes::all::{OP_CHECKSIGVERIFY, OP_CSV, OP_EQUALVERIFY, OP_SHA256, OP_SIZE},
+    opcodes::all::{OP_EQUALVERIFY, OP_SHA256, OP_SIZE},
     psbt::Input,
-    relative,
     taproot::{ControlBlock, LeafVersion},
     Address, Network, ScriptBuf,
 };
-use secp256k1::{schnorr::Signature, XOnlyPublicKey};
+use secp256k1::XOnlyPublicKey;
 use strata_bridge_primitives::scripts::prelude::*;
 
-use crate::connectors::witness_data::WitnessData;
+use crate::witness_data::WitnessData;
 
-/// The connector to move the operator's stake across transactions.
-// TODO: Replace this with `ConnectorStake`.
-#[derive(Debug, Clone, Copy)]
-pub struct ConnectorS {
-    /// The N-of-N aggregated public key for the operator set.
-    n_of_n_agg_pubkey: XOnlyPublicKey,
-
-    /// The bitcoin network on which the connector operates.
-    network: Network,
-}
-
-impl ConnectorS {
-    /// Creates a new `ConnectorS` with the given N-of-N aggregated public key and the
-    /// bitcoin network.
-    pub fn new(n_of_n_agg_pubkey: XOnlyPublicKey, network: Network) -> Self {
-        Self {
-            n_of_n_agg_pubkey,
-            network,
-        }
-    }
-
-    /// Creates a taproot address with key spend path for the given operator set.
-    pub fn create_taproot_address(&self) -> Address {
-        let (addr, _spend_info) = create_taproot_addr(
-            &self.network,
-            SpendPath::KeySpend {
-                internal_key: self.n_of_n_agg_pubkey,
-            },
-        )
-        .expect("must be able to create taproot address");
-
-        addr
-    }
-
-    /// Finalizes a psbt input where this connector is used with the provided signature.
-    ///
-    /// # Note
-    ///
-    /// This method does not check if the signature is valid for the input. It is the caller's
-    /// responsibility to ensure that the signature is valid.
-    ///
-    /// If the psbt input is already in the final state, then this method overrides the signature.
-    pub fn create_tx_input(&self, signature: Signature, input: &mut Input) {
-        finalize_input(input, [signature.as_ref()]);
-    }
-}
-
-/// The connector to move the operator's stake across Stake transactions.
+/// The connector to decide whether the operator's stake can be used for a withdrawal (Payout
+/// Optimistic) or not (Burn Payouts).
 ///
-/// It is used in the Disprove and Slash Stake `k` transactions, where `k` is the index of the
-/// stake transaction.
-///
-/// The operator can also advance the stake chain by revealing the preimage, along with a valid
-/// signature from the operator's public key.
-/// Note that the stake advancement is done by the `stake-chain` crate.
+/// It is used in the Payout Optimistic and Burn Payouts transactions.
 ///
 /// To illustrate the concept, let's say that an operator wants to claim the `k`th bridged-in UTXO.
 /// For this, they need the `k`th Claim Transaction and hence the `k`th stake transaction. An
@@ -81,23 +32,18 @@ impl ConnectorS {
 ///
 /// If the operator has received a `k`th Payout Optimistic or Payout transaction, they can advance
 /// the stake chain (revealing the preimage) without fear. It is the responsibility of the
-/// [`ConnectorP`](super::connector_p::ConnectorP) and [`ConnectorStake`] to ensure that will make
-/// it impossible make it impossible
+/// [`ConnectorP`] and [`ConnectorStake`](super::connector_s::ConnectorStake) to ensure that will
+/// make it impossible
 ///
 /// # Security
 ///
-/// An operator can only advance the stake chain if they reveal the preimage along with a valid
-/// signature from the operator's public key. Hence, the operator must must be able to provide the
-/// preimage to the [`ConnectorStake`]. It is required that the preimage be securely derived and
-/// never reused under any circumstances
-// TODO: This should replace the `ConnectorS` struct above.
+/// An operator can only advance the stake chain if they reveal the preimage. Hence, the operator
+/// must be able to provide the preimage to the [`ConnectorP`]. It is required that the preimage be
+/// securely derived and never reused under any circumstances.
 #[derive(Debug, Clone, Copy)]
-pub struct ConnectorStake {
+pub struct ConnectorP {
     /// The N-of-N aggregated public key for the operator set.
     n_of_n_agg_pubkey: XOnlyPublicKey,
-
-    /// The operator's public key.
-    operator_pubkey: XOnlyPublicKey,
 
     /// The hash of the `k`th stake preimage.
     ///
@@ -106,28 +52,21 @@ pub struct ConnectorStake {
     /// validating transactions before operators offer up their signatures.
     stake_hash: sha256::Hash,
 
-    /// The `ΔS` interval relative timelock to advance the stake chain.
-    delta: relative::LockTime,
-
     /// The bitcoin network on which the connector operates.
     network: Network,
 }
 
-impl ConnectorStake {
-    /// Creates a new [`ConnectorStake`] with the given N-of-N aggregated public key, `k`th stake
+impl ConnectorP {
+    /// Creates a new [`ConnectorP`] with the given N-of-N aggregated public key, `k`th stake
     /// preimage, and the bitcoin network.
     pub fn new(
         n_of_n_agg_pubkey: XOnlyPublicKey,
-        operator_pubkey: XOnlyPublicKey,
         stake_hash: sha256::Hash,
-        delta: relative::LockTime,
         network: Network,
     ) -> Self {
         Self {
             n_of_n_agg_pubkey,
-            operator_pubkey,
             stake_hash,
-            delta,
             network,
         }
     }
@@ -139,35 +78,31 @@ impl ConnectorStake {
     /// The locking script can be represented as the following miniscript policy:
     ///
     /// ```text
-    /// thresh(3,pk(operator_pubkey), sha256(stake_preimage), older(ΔS))
+    /// sha256(stake_preimage)
     /// ```
     ///
     /// which compiles to the following script:
     ///
     /// ```text
-    /// <operator_pubkey> OP_CHECKSIGVERIFY OP_SIZE <20> OP_EQUALVERIFY OP_SHA256
-    /// <stake_preimage> OP_EQUALVERIFY <ΔS> OP_CHECKSEQUENCEVERIFY
+    /// OP_SIZE <20> OP_EQUALVERIFY OP_SHA256 <stake_preimage> OP_EQUALVERIFY
     /// ```
     pub fn generate_script(&self) -> ScriptBuf {
         ScriptBuf::builder()
-            .push_slice(self.operator_pubkey.serialize())
-            .push_opcode(OP_CHECKSIGVERIFY)
             .push_opcode(OP_SIZE)
             .push_int(0x20)
             .push_opcode(OP_EQUALVERIFY)
             .push_opcode(OP_SHA256)
             .push_slice(self.stake_hash.to_byte_array())
             .push_opcode(OP_EQUALVERIFY)
-            .push_sequence(self.delta.into())
-            .push_opcode(OP_CSV)
+            .push_int(1)
             .into_script()
     }
 
     /// Creates a P2TR address with key spend path for the given operator set and a single script
-    /// path that can be unlocked by revealing the preimage, along with an operator signature and
-    /// is timelocked by `ΔS`.
+    /// path that can be unlocked by revealing the preimage.
     ///
-    /// This is used to advance the stake chain, slash the stake, and disprove the stake.
+    /// This is used to invalidate that a certain stake can be used in payouts by an operator in a
+    /// Burn Payouts transaction.
     ///
     /// See [`Self::generate_script`] for the script implementation details.
     pub fn generate_address(&self) -> Address {
@@ -219,14 +154,9 @@ impl ConnectorStake {
             WitnessData::Signature(signature) => {
                 finalize_input(input, [&signature.serialize().to_vec()]);
             }
-            WitnessData::Both {
-                signature,
-                preimage,
-            } => finalize_input(
-                input,
-                // NOTE: Order matters here.
-                vec![&preimage.to_vec(), &signature.serialize().to_vec()],
-            ),
+            WitnessData::Preimage(preimage) => {
+                finalize_input(input, [&preimage]);
+            }
             _ => (), // other variants are no-op.
         }
     }
@@ -235,24 +165,20 @@ impl ConnectorStake {
 #[cfg(test)]
 mod tests {
     use bitcoin::{
-        absolute, consensus,
-        hashes::Hash,
-        sighash::{self, Prevouts, SighashCache},
-        taproot::LeafVersion,
-        transaction, Amount, BlockHash, OutPoint, TapLeafHash, Transaction, TxIn, TxOut, Witness,
+        absolute, consensus, transaction, Amount, BlockHash, OutPoint, Transaction, TxIn, TxOut,
+        Witness,
     };
     use corepc_node::{serde_json::json, Conf, Node};
-    use secp256k1::{Message, SECP256K1};
     use strata_bridge_test_utils::prelude::generate_keypair;
     use strata_btcio::rpc::types::SignRawTransactionWithWallet;
     use strata_common::logging::{self, LoggerConfig};
-    use tracing::{info, trace};
+    use tracing::info;
 
     use super::*;
 
     #[test]
-    fn connector_s_script_path() {
-        logging::init(LoggerConfig::new("connector-s-script-path".to_string()));
+    fn connector_p_script_path() {
+        logging::init(LoggerConfig::new("connector-p-script-path".to_string()));
 
         // Setup Bitcoin node
         let mut conf = Conf::default();
@@ -285,25 +211,14 @@ mod tests {
             .expect("must be able to get the coinbase transaction")
             .compute_txid();
 
-        // Generate keys
+        // Generate keys and preimage
         let n_of_n_keypair = generate_keypair();
-        let operator_keypair = generate_keypair();
         let n_of_n_pubkey = n_of_n_keypair.x_only_public_key().0;
-        let operator_pubkey = operator_keypair.x_only_public_key().0;
-
-        // Generate stake preimage
-        let stake_preimage = [1; 32];
+        let stake_preimage = [1u8; 32];
         let stake_hash = sha256::Hash::hash(&stake_preimage);
 
-        // Create relative timelock (e.g., 10 blocks)
-        let delta = relative::LockTime::from_height(10);
-
         // Create connector
-        let connector_s =
-            ConnectorStake::new(n_of_n_pubkey, operator_pubkey, stake_hash, delta, network);
-
-        // Generate address and script
-        let taproot_script = connector_s.generate_address().script_pubkey();
+        let connector_p = ConnectorP::new(n_of_n_pubkey, stake_hash, network);
 
         // Create funding transaction
         let funding_input = OutPoint {
@@ -324,7 +239,7 @@ mod tests {
         let output = vec![
             TxOut {
                 value: funding_amount,
-                script_pubkey: taproot_script.clone(),
+                script_pubkey: connector_p.generate_address().script_pubkey(),
             },
             TxOut {
                 value: coinbase_amount
@@ -343,7 +258,7 @@ mod tests {
             output,
         };
 
-        // Sign the transaction
+        // Sign and broadcast funding transaction
         let signed_funding_tx = btc_client
             .call::<SignRawTransactionWithWallet>(
                 "signrawtransactionwithwallet",
@@ -355,7 +270,6 @@ mod tests {
         let signed_funding_tx =
             consensus::encode::deserialize_hex(&signed_funding_tx.hex).expect("must deserialize");
 
-        // Broadcast the funding transaction
         let funding_txid = btc_client
             .send_raw_transaction(&signed_funding_tx)
             .expect("must be able to broadcast transaction")
@@ -364,12 +278,12 @@ mod tests {
 
         info!(%funding_txid, "Funding transaction broadcasted");
 
-        // Mine the funding transaction with sufficient blocks for the relative timelock
+        // Mine the funding transaction
         let _ = btc_client
-            .generate_to_address((delta.to_consensus_u32() as usize) + 1, &funded_address)
+            .generate_to_address(1, &funded_address)
             .expect("must be able to generate blocks");
 
-        // Create the transaction that spents the connector s
+        // Create spending transaction that spends the connector p
         let spending_input = OutPoint {
             txid: funding_txid,
             vout: 0,
@@ -385,52 +299,27 @@ mod tests {
             lock_time: absolute::LockTime::ZERO,
             input: vec![TxIn {
                 previous_output: spending_input,
-                sequence: delta.into(), // Important: Set the sequence number to match the timelock
                 ..Default::default()
             }],
             output: vec![spending_output],
         };
 
-        // Create sighash for the spending transaction
-        let mut sighash_cache = SighashCache::new(&spending_tx);
-        let sighash_type = sighash::TapSighashType::Default;
-        // Create the prevouts
-        let prevouts = [TxOut {
-            value: funding_amount,
-            script_pubkey: taproot_script,
-        }];
-        let prevouts = Prevouts::All(&prevouts);
-
         // Create the locking script
-        let locking_script = connector_s.generate_script();
+        let locking_script = connector_p.generate_script();
 
         // Get taproot spend info
-        let (_, control_block) = connector_s.generate_spend_info();
-
-        let leaf_hash =
-            TapLeafHash::from_script(locking_script.as_script(), LeafVersion::TapScript);
-        let sighash = sighash_cache
-            .taproot_script_spend_signature_hash(0, &prevouts, leaf_hash, sighash_type)
-            .expect("must create sighash");
-
-        let message =
-            Message::from_digest_slice(sighash.as_byte_array()).expect("must create a message");
-
-        // Sign the transaction with operator key
-        let signature = SECP256K1.sign_schnorr(&message, &operator_keypair);
-        trace!(%signature, "Signature");
+        let (_, control_block) = connector_p.generate_spend_info();
 
         // Construct the witness stack
         let mut witness = Witness::new();
         witness.push(stake_preimage);
-        witness.push(signature.as_ref());
         witness.push(locking_script.to_bytes());
         witness.push(control_block.serialize());
 
         // Set the witness in the transaction
         spending_tx.input[0].witness = witness;
 
-        // Try to broadcast the spending transaction
+        // Broadcast spending transaction
         let spending_txid = btc_client
             .send_raw_transaction(&spending_tx)
             .expect("must be able to broadcast spending transaction")
