+++ crates/connectors/src/partial_verification_scripts.rs
@@ -14,6 +14,12 @@ pub static PARTIAL_VERIFIER_SCRIPTS: LazyLock<[Script; 381]> =
     LazyLock::new(load_or_create_verifier_scripts);
 
 /// Loads tapscripts for the groth16 verifier program.
+///
+/// NOTE: these scripts must be regenerated whenever the following change:
+///
+/// * The verification key for the groth16 verifier program.
+/// * The number and nature of tapscripts generated by bitvm.
+/// * The structure of the verifier program (i.e., the proof statements/guest build).
 pub fn load_or_create_verifier_scripts() -> [Script; 381] {
     if cfg!(feature = "mock") {
         warn!("Detected mock feature, returning empty verifier scripts");
