+++ crates/connectors/src/prelude.rs
@@ -1,3 +1,4 @@
+//! This module exports all the connectors in this crate for convenience.
 pub use super::{
     connector_a30::*, connector_a31::*, connector_c0::*, connector_c1::*, connector_cpfp::*,
     connector_k::*, connector_p::*, connector_s::*, connectors_a::*, witness_data::*,
