+++ crates/tx-graph/src/lib.rs
@@ -5,8 +5,6 @@
 #![feature(generic_const_exprs)] // this feature is used in size computations
 #![feature(array_try_from_fn)] // this feature is used to generate arrays in a fallible way
 
-pub mod connectors;
 pub mod errors;
-pub mod partial_verification_scripts;
 pub mod peg_out_graph;
 pub mod transactions;
