+++ crates/tx-graph/src/transactions/assert_chain.rs
@@ -1,9 +1,9 @@
 use bitcoin::Txid;
+use strata_bridge_connectors::prelude::*;
 use strata_bridge_primitives::{params::connectors::*, types::OperatorIdx};
 use tracing::trace;
 
 use super::prelude::*;
-use crate::connectors::prelude::*;
 
 /// Data needed to construct an [`AssertChain`].
 #[derive(Debug, Clone)]
